pyproject.toml
# プロジェクト設定ファイル

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "createPPT"
version = "0.1.0"
description = "PowerPoint自動生成アプリケーション"
requires-python = ">=3.11"
dependencies = [
    "python-pptx>=0.6.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.coverage.run]
source = ["src"]
